+++ src/client.js
@@ -1,7 +1,6 @@
 const axios = require("axios").default;
 const errors = require("./errors");
 const EventEmitter = require("events");
-const { WebSocket } = require("./websocket");
 const pkg = require("../package.json");
 
 exports.Client = class Client extends EventEmitter {
@@ -18,24 +17,13 @@ exports.Client = class Client extends EventEmitter {
          * The payouts API Url
          * @type {String}
          */
-        this.API_URL_PAYOUTS = "https://api.payouts.unmineable.com";
+        this.API_URL_PAYOUTS = "https://payouts.unmineable.com";
 
         /**
          * The WebSocket Url
          * @type {String}
          */
         this.WEBSOCKET_URL = "wss://ws.unminable.com";
-    }
-
-    /**
-     * Build url to the API
-     * @param  {String} version API version
-     * @param  {String} path API endpoint
-     * @return {String} url
-     * @private
-     */
-    #buildURL(version, path) {
-        return `${this.API_URL}/${version}/${path}`;
     }
 
     /**
@@ -49,11 +37,9 @@ exports.Client = class Client extends EventEmitter {
     #request(method, version, path, options = {}) {
         Object.assign(options, {
             method: method,
-            url: options.url ? options.url : this.#buildURL(version, path),
+            url: options.url ? options.url : `${this.API_URL}/${version}/${path}`,
             headers: {
-                "User-Agent": `unmineable.js ${
-                    pkg.version
-                } (https://github.com/LockBlock-dev/unmineable.js)`,
+                "User-Agent": `unmineable.js ${pkg.version} (https://github.com/LockBlock-dev/unmineable.js)`,
                 "Content-Type": "application/json",
                 "Accept-Encoding": "UTF8",
             },
@@ -100,49 +86,111 @@ exports.Client = class Client extends EventEmitter {
      * @example start("0cc5691e-cedb-11eb-b8bc-0242ac130003")
      */
     start(uuid) {
-        const ws = new WebSocket(this.WEBSOCKET_URL);
+        const socket = io(this.WEBSOCKET_URL, { path: "/", transports: ["websocket"] });
 
-        ws.on("worker", (data) => {
-            this.emit("worker", data);
+        socket.on("connect", () => {
+            socket.emit("join_account", uuid);
         });
 
-        ws.on("total_paid", (data) => {
-            this.emit("total_paid", data);
+        socket.on("disconnect", (reason) => {
+            if (reason === "io server disconnect") {
+                // the disconnection was initiated by the server, you need to reconnect manually
+                socket.connect();
+            }
+            // else the socket will automatically try to reconnect
         });
 
-        ws.on("total_workers", (data) => {
-            this.emit("total_workers", data);
+        socket.on("hr_worker", (data) => {
+            this.emit("worker", {
+                name: data.n,
+                reportedHashrate: data.rh,
+                calculatedHashrate: data.ch,
+                lastShareAt: data.b,
+                online: data.o,
+            });
         });
 
-        ws.on("payment_status", (data) => {
+        socket.on("stats_paid", (data) => {
+            /**
+             * Emits the total paid by the pool in BTC.
+             * @event Client#total_paid
+             * @type {Object}
+             * @property {number} total BTC total
+             */
+            this.emit("total_paid", { total: parseFloat(data) });
+        });
+
+        socket.on("stats_workers", (data) => {
+            /**
+             * Emits the total of active workers on the pool.
+             * @event Client#total_workers
+             * @type {Object}
+             * @property {number} total BTC total
+             */
+            this.emit("total_workers", { total: data });
+        });
+
+        socket.on("payment_status", (data) => {
             this.emit("payment_status", data);
         });
 
-        ws.on("reward", (data) => {
+        socket.on("reward", (data) => {
             this.emit("reward", data);
         });
 
-        ws.on("refresh", (data) => {
+        socket.on("refresh", (data) => {
             this.emit("refresh", data);
         });
 
-        ws.on("ethash_last_reward", (data) => {
-            this.emit("ethash_last_reward", data);
+        socket.on("stats_ethash_last_reward", (data) => {
+            /**
+             * Emits when the pool gets a new ethash block reward.
+             * @event Client#ethash_last_reward
+             * @type {Object}
+             * @property {number} timestamp
+             */
+            this.emit("ethash_last_reward", { timestamp: data });
         });
 
-        ws.on("etchash_last_reward", (data) => {
-            this.emit("etchash_last_reward", data);
+        socket.on("stats_etchash_last_reward", (data) => {
+            /**
+             * Emits when the pool gets a new etchash block reward.
+             * @event Client#etchash_last_reward
+             * @type {Object}
+             * @property {number} timestamp
+             */
+            this.emit("etchash_last_reward", { timestamp: data });
         });
 
-        ws.on("randomx_last_reward", (data) => {
-            this.emit("randomx_last_reward", data);
+        socket.on("stats_randomx_last_reward", (data) => {
+            /**
+             * Emits when the pool gets a new randomx block reward.
+             * @event Client#randomx_last_reward
+             * @type {Object}
+             * @property {number} timestamp
+             */
+            this.emit("randomx_last_reward", { timestamp: data });
         });
 
-        ws.on("x16rv2_last_reward", (data) => {
-            this.emit("x16rv2_last_reward", data);
+        socket.on("stats_x16rv2_last_reward", (data) => {
+            /**
+             * Emits when the pool gets a new RVN block reward.
+             * @event Client#x16rv2_last_reward
+             * @type {Object}
+             * @property {number} timestamp
+             */
+            this.emit("x16rv2_last_reward", { timestamp: data });
         });
 
-        ws.start(uuid);
+        socket.on("stats_autolykos_last_reward", (data) => {
+            /**
+             * Emits when the pool gets a new autolykos block reward.
+             * @event Client#autolykos_last_reward
+             * @type {Object}
+             * @property {number} timestamp
+             */
+            this.emit("autolykos_last_reward", { timestamp: data });
+        });
     }
 
     //----- WEB API ------
@@ -187,7 +235,7 @@ exports.Client = class Client extends EventEmitter {
      */
     version(params = {}) {
         if (params.type == "website") return this.#request("GET", "v3", "site/version");
-        if (params.type == "miner") return this.#request("GET", "v2", "miner/version");
+        if (params.type == "miner") return this.#request("GET", "v3", "miner/version");
     }
 
     /**
@@ -200,7 +248,7 @@ exports.Client = class Client extends EventEmitter {
     }
 
     /**
-     * Get a wallet | account stats.
+     * Get a wallet | account info.
      * @param {Object} params query params
      * @param {String} [params.address] wallet address
      * @param {String} [params.uuid] uuid
@@ -213,6 +261,16 @@ exports.Client = class Client extends EventEmitter {
         if (params.uuid) return this.#request("GET", "v4", `account/${params.uuid}`);
         if (params.address)
             return this.#request("GET", "v4", `address/${params.address}?coin=${params.coin}`);
+    }
+
+    /**
+     * Get an account stats.
+     * @param {String} uuid uuid
+     * @example accountStats("0cc5691e-cedb-11eb-b8bc-0242ac130003")
+     * @return {Promise<Object>} promise
+     */
+    accountStats(uuid) {
+        return this.#request("GET", "v4", `account/${uuid}/stats`);
     }
 
     /**
@@ -254,8 +312,9 @@ exports.Client = class Client extends EventEmitter {
      * @param {String} body.algo algorithms:
      * - ethash
      * - etchash
-     * - x16rv2 (kawpow)
+     * - kawpow
      * - randomx
+     * - autolykos
      * @param {String} body.coin coin symbol
      * @param {Number} body.hashrate hashrate for this algorithm
      * - ethash, etchash, x16rv2: MH/s
@@ -274,22 +333,22 @@ exports.Client = class Client extends EventEmitter {
      * @param {Object} body request body
      * @param {String} body.uuid uuid
      * @param {Boolean} [body.autoPay=false] auto pay setting
-     * @param {String} [body.network] network setting
+     * @param {String} body.network network setting
      * - ETH
      * - BSC
+     * - ARBITRUM
      * @example settings({ uuid: "0cc5691e-cedb-11eb-b8bc-0242ac130003", autoPay: true })
      * settings({ uuid: "0cc5691e-cedb-11eb-b8bc-0242ac130003", network: "BSC" })
      * @return {Promise<Object>} promise
      */
     settings(body = {}) {
-        if (body.autoPay !== undefined) {
+        if (body.autoPay) {
             return this.#request("POST", "v1", `address/${body.uuid}/setting/auto_pay`, {
                 data: { setting: body.autoPay },
             });
         } else if (body.network) {
             return this.#request("POST", null, null, {
                 url: `${this.API_URL_PAYOUTS}/v3/address/${body.uuid}/switch?network=${body.network}`,
-                data: { setting: body.autoPay },
             });
         }
     }