+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2022-09-14
+
+### Added
+
+-   accountStats method (differs from wallet method)
+-   JSDoc for some websocket events
+-   Autolykos reward websocket event
+
+### Changed
+
+-   fixed API_URL_PAYOUTS url
+-   websocket.js code is now directly in client.js
+-   miner version now uses v3 API
+-   dependencies update
+-   License date
+
+### Removed
+
+-   buildURL useless method
+-   payment_status event console.log
+-   autoPay setting when changing network
+
 ## [2.0.0] - 2021-07-28
 
 ### Added